Fix glitch that lead to running one sample only
Fixed a glitch in which mapping and downstream processes
run for only one sample. That is because output from index process is
one path and it is treated as que channel that will be exhausted after
one time run, and the precess will run once. The solution is to convert
the output from index process to value channel by using
"bowtie2_index.out.first()" as input for mapping process.